requirements.txt
nicegui>=1.4.0
python-dotenv>=1.0.0
httpx>=0.24.1
pydantic>=2.0.0 